Adventurer actions methods
added methods for Fuze's actions and created inventory array for the
items.

Have to figure out some things with having the name of items written out
when it is dropped or picked up or used.

I also called the class to the main class in Finding Milo. Have to see
in src file if this will work.